+++ src/Controller/PlanMensualGeneralController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\ActividadGeneral;
+use App\Entity\MiembroConsejoDireccion;
 use App\Entity\PlanMensualGeneral;
 use App\Form\PlanMensualGeneralType;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
@@ -22,21 +23,28 @@ class PlanMensualGeneralController extends Controller
      */
     public function index(Request $request): Response
     {
-        $planmensualgenerals = $this->getDoctrine()->getRepository(PlanMensualGeneral::class)->findAll();
+        $em=$this->getDoctrine()->getManager();
+        $esmiembro = null!=$em->getRepository(MiembroConsejoDireccion::class)->findOneByUsuario($this->getUser());
 
+        if(!$esmiembro && !$this->isGranted('ROLE_COORDINADOR'))
+            throw $this->createAccessDeniedException();
+
+        $planmensualgenerals = $this->getDoctrine()->getRepository(PlanMensualGeneral::class)->findAll();
+        $esCoordinador=$this->isGranted('ROLE_COORDINADOR');
         if ($request->isXmlHttpRequest())
             return $this->render('planmensualgeneral/_table.html.twig', [
                 'planmensualgenerals' => $planmensualgenerals,
+                'esCoordinador'=>$esCoordinador
             ]);
 
-        $parameters = [
+        return $this->render('planmensualgeneral/index.html.twig', [
             'user_id' => $this->getUser()->getId(),
             'user_foto'=>null!=$this->getUser()->getRutaFoto() ? $this->getUser()->getRutaFoto() : null,
             'user_nombre'=>$this->getUser()->getNombre(),
             'user_correo'=>$this->getUser()->getCorreo(),
-            'planmensualgenerals' => $planmensualgenerals];
-
-        return $this->render('planmensualgeneral/index.html.twig', $parameters);
+            'planmensualgenerals' => $planmensualgenerals,
+            'esCoordinador'=>$esCoordinador
+        ]);
     }
 
     /**
@@ -44,6 +52,10 @@ class PlanMensualGeneralController extends Controller
      */
     public function new(Request $request): Response
     {
+        if (!$request->isXmlHttpRequest())
+            throw $this->createAccessDeniedException();
+
+        $this->denyAccessUnlessGranted('ROLE_COORDINADOR');
         $planmensualgeneral = new PlanMensualGeneral();
         $em = $this->getDoctrine()->getManager();
         $form = $this->createForm(PlanMensualGeneralType::class, $planmensualgeneral, array('action' => $this->generateUrl('planmensualgeneral_new')));
@@ -56,6 +68,8 @@ class PlanMensualGeneralController extends Controller
                 return new JsonResponse(array('mensaje' => 'El plan mensual fue registrado satisfactoriamente',
                     'mes' => $planmensualgeneral->getMestoString(),
                     'anno' => $planmensualgeneral->getAnno(),
+                    'fechainicio'=>$planmensualgeneral->getEdicionfechainicio()->format('d-m-Y'),
+                    'fechafin'=>$planmensualgeneral->getEdicionfechafin()->format('d-m-Y'),
                     'id' => $planmensualgeneral->getId(),
                 ));
             } else {
@@ -77,16 +91,65 @@ class PlanMensualGeneralController extends Controller
      */
     public function show(Request $request, PlanMensualGeneral $planmensualgeneral): Response
     {
+        $this->denyAccessUnlessGranted('VIEW',$planmensualgeneral);
         $actividads = $this->getDoctrine()
             ->getRepository(ActividadGeneral::class)
             ->findBy(array('planmensualgeneral' => $planmensualgeneral));
 
+        $today=new \DateTime('today');
+        $enTiempo=$today>=$planmensualgeneral->getEdicionfechainicio() && $today<= $planmensualgeneral->getEdicionfechafin();
+        dump($today);
+        dump($planmensualgeneral->getEdicionfechafin());
+        dump($enTiempo);
         return $this->render('planmensualgeneral/show.html.twig', ['planmensualgeneral' => $planmensualgeneral,
             'actividads' => $actividads,
             'user_id' => $this->getUser()->getId(),
             'user_foto'=>null!=$this->getUser()->getRutaFoto() ? $this->getUser()->getRutaFoto() : null,
             'user_nombre'=>$this->getUser()->getNombre(),
             'user_correo'=>$this->getUser()->getCorreo(),
+            'enTiempo'=>$enTiempo
+        ]);
+    }
+
+    /**
+     * @Route("/{id}/edit", name="planmensualgeneral_edit",options={"expose"=true}, methods="GET|POST")
+     */
+    public function edit(Request $request, PlanMensualGeneral $plan): Response
+    {
+        if (!$request->isXmlHttpRequest())
+            throw $this->createAccessDeniedException();
+
+        $this->denyAccessUnlessGranted('EDIT',$plan);
+        $form = $this->createForm(PlanMensualGeneralType::class, $plan,
+            array('action' => $this->generateUrl('planmensualgeneral_edit', array('id' => $plan->getId()))));
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted())
+            if ($form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($plan);
+                $em->flush();
+                return new JsonResponse(array('mensaje' => 'El plan mensual fue actualizado satisfactoriamente',
+                    'mes' => $plan->getMestoString(),
+                    'anno' => $plan->getAnno(),
+                    'fechainicio'=>$plan->getEdicionfechainicio()->format('d-m-Y'),
+                    'fechainicio'=>$plan->getEdicionfechafin()->format('d-m-Y'),
+                ));
+            } else {
+                $page = $this->renderView('planmensualgeneral/_form.html.twig', array(
+                    'form' => $form->createView(),
+                    'form_id' => 'planmensualgeneral_edit',
+                    'action' => 'Actualizar',
+                ));
+                return new JsonResponse(array('form' => $page, 'error' => true));
+            }
+
+        return $this->render('planmensualgeneral/_new.html.twig', [
+            'plan' => $plan,
+            'title' => 'Editar plan mensual',
+            'action' => 'Actualizar',
+            'form_id' => 'planmensualgeneral_edit',
+            'form' => $form->createView(),
         ]);
     }
 
@@ -98,6 +161,7 @@ class PlanMensualGeneralController extends Controller
         if (!$request->isXmlHttpRequest())
             throw $this->createAccessDeniedException();
 
+        $this->denyAccessUnlessGranted('ROLE_COORDINADOR');
         $em = $this->getDoctrine()->getManager();
         $em->remove($planmensualgeneral);
         $em->flush();
